Refactor events API calls to do everything in-sync.
It turned out to be really difficult and unreliable to do the async JS
method. This is more reliable. You'll need a server with a fast connection
to SiftScience to avoid lags in your website though.

That said, I may revisit the async issue.
There is a promising wp-async library out there.